+++ Hydra/Hydra/Terrain/Marching/Marching.h
@@ -13,7 +13,7 @@ namespace Hydra
 	public:
 		Marching(float surfaceThreshold = 0.5f);
 
-		Mesh* Generate(float* voxels, int width, int height, int depth);
+		Mesh* Generate(int* voxels, int width, int height, int depth);
 
 	protected:
 		virtual void March(float x, float y, float z, float* cube, Mesh* mesh) = 0;
